Ensure we can load check definitions inside a profile
